requirements.txt
beautifulsoup4==4.11.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
colorama==0.4.4
idna==3.3
python-dotenv==0.20.0
requests==2.28.0
soupsieve==2.3.2.post1
urllib3==1.26.9
wikipedia==1.4.0
